feat: 완료된 업무 아이템에 완료 일자 표기
